requirements.txt
google-cloud-storage
google-cloud-aiplatform
pandas